Fix multiple operator input bug
If the user input num1, an operator, num2, another operator, and then
another operator after that, a bug in the operator chaining logic
caused it to evaluate the expression with an empty num2. Change the
boolean expression to check for num2 being an empty string instead of
checking the mode.